Add Consumer.ReleaseOK() method
This allows you to short circuit the Synchronization.DwellTime timeout
during a rebalance if you know that you have already finished processing
all in-progress messages.

ReleaseOK() returns a chan you write to to indicate that it is ok to
short circuit the DwellTimeout. This allows for the caller to timeout in
cases where the DwellTime timer might have already fired.